+++ cpp/hello/tests/test_callback.cpp
@@ -2,28 +2,28 @@
 
 #include <callback.hpp>
 
-float mapper(float item)
+int mapper(int item)
 {
     return item * 2;
 }
 
 TEST_CASE("[C] hello__callback__map_safe()")
 {
-    float items_in[] = {1.f, 2.f, 3.f};
+    int items_in[] = {1, 2, 3};
     const size_t item_count = sizeof(items_in) / sizeof(items_in[0]);
 
-    float items_out[item_count];
+    int items_out[item_count];
 
     hello__callback__map_safe(items_in, items_out, item_count, mapper);
 
-    REQUIRE(items_out[0] == 2.f);
-    REQUIRE(items_out[1] == 4.f);
-    REQUIRE(items_out[2] == 6.f);
+    REQUIRE(items_out[0] == 2);
+    REQUIRE(items_out[1] == 4);
+    REQUIRE(items_out[2] == 6);
 }
 
 TEST_CASE("[C++] hello::callback::map()")
 {
-    auto result = hello::callback::map({1.f, 2.f, 3.f}, [](float item) { return item * 2; });
+    auto result = hello::callback::map({1, 2, 3}, [](int item) { return item * 2; });
 
-    REQUIRE(result == std::vector<float>{2.f, 4.f, 6.f});
+    REQUIRE(result == std::vector<int>{2, 4, 6});
 }
